requirements.txt
PySide6~=6.9.0
Pypubsub~=4.0.3
Markdown~=3.7
SQLAlchemy~=2.0.40
mistune~=3.1.3
minio~=7.2.15
oss2~=2.19.1
requests~=2.32.3
beautifulsoup4~=4.13.3
pypandoc~=1.15
dotenv~=0.9.9
python-dotenv~=1.1.0
psutil